feat: render products
