+++ src/components/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import MainPage from "./pages/MainPage";
 import SignUpPage from "./pages/SignUpPage";
-import HomePage from "./pages/HomePage";
+import HomePage from "./pages/Homepage/HomePage";
 import Cart from "./pages/Cart";
 import Checkout from "./pages/Checkout";
 import UserContext from "../contexts/UserContext";
